+++ firmware/ftl/include/ftl/functional.hpp
